+++ features/admin.actions.v1/pages/action-configuration-page.tsx
@@ -51,10 +51,13 @@ import useGetActionById from "../api/use-get-action-by-id";
 import useGetActionsByType from "../api/use-get-actions-by-type";
 import PreIssueAccessTokenActionConfigForm from "../components/pre-issue-access-token-action-config-form";
 import PreUpdatePasswordActionConfigForm from "../components/pre-update-password-action-config-form";
+import PreUpdateProfileActionConfigForm from "../components/pre-update-profile-action-config-form";
 import { ActionsConstants } from "../constants/actions-constants";
 import {
     ActionConfigFormPropertyInterface, PreUpdatePasswordActionConfigFormPropertyInterface,
-    PreUpdatePasswordActionResponseInterface
+    PreUpdatePasswordActionResponseInterface,
+    PreUpdateProfileActionConfigFormPropertyInterface,
+    PreUpdateProfileActionResponseInterface
 } from "../models/actions";
 import "./action-configuration-page.scss";
 import { useHandleError, useHandleSuccess } from "../util/alert-util";
@@ -154,6 +157,19 @@ const ActionConfigurationPage: FunctionComponent<ActionConfigurationPageInterfac
                     certificate: (action as PreUpdatePasswordActionResponseInterface)?.passwordSharing.certificate
                         || "",
                     passwordSharing: (action as PreUpdatePasswordActionResponseInterface)?.passwordSharing.format
+                };
+            } else {
+                return null;
+            }
+        }, [ action ]);
+
+    const preUpdateProfileActionInitialValues: PreUpdateProfileActionConfigFormPropertyInterface =
+        useMemo(() => {
+
+            if (action && actionTypeApiPath === ActionsConstants.PRE_UPDATE_PROFILE_API_PATH ) {
+                return {
+                    ...actionCommonInitialValues,
+                    attributes: (action as PreUpdateProfileActionResponseInterface)?.attributes
                 };
             } else {
                 return null;
@@ -407,6 +423,16 @@ const ActionConfigurationPage: FunctionComponent<ActionConfigurationPageInterfac
                             { actionTypeApiPath === ActionsConstants.PRE_UPDATE_PASSWORD_API_PATH && (
                                 <PreUpdatePasswordActionConfigForm
                                     initialValues={ preUpdatePasswordActionInitialValues }
+                                    isLoading={ isLoading }
+                                    isReadOnly={ isReadOnly() }
+                                    actionTypeApiPath={ actionTypeApiPath }
+                                    isCreateFormState={ showCreateForm }
+                                />
+                            )
+                            }
+                            { actionTypeApiPath === ActionsConstants.PRE_UPDATE_PROFILE_API_PATH && (
+                                <PreUpdateProfileActionConfigForm
+                                    initialValues={ preUpdateProfileActionInitialValues }
                                     isLoading={ isLoading }
                                     isReadOnly={ isReadOnly() }
                                     actionTypeApiPath={ actionTypeApiPath }