+++ features/admin.registration-flow-builder.v1/api/use-get-supported-profile-attributes.ts
@@ -47,6 +47,9 @@ const useGetSupportedProfileAttributes = <Data = Attribute[], Error = RequestErr
             "Content-Type": "application/json"
         },
         method: HttpMethods.GET,
+        params: {
+            "exclude-identity-claims": true
+        },
         url: store.getState().config.endpoints.localClaims
     };