+++ modules/i18n/src/models/namespaces/actions-ns.ts
@@ -184,7 +184,15 @@ export interface actionsNS {
                 }
             };
             label: string;
-        }
+        };
+        userAttributes: {
+            heading: string;
+            hint: string;
+            search: {
+                placeholder: string;
+                clearButton: string;
+            };
+        };
     };
     goBackActions: string,
     notification: {
@@ -279,6 +287,12 @@ export interface actionsNS {
                     message : string;
                 };
                 delete: {
+                    description: string;
+                    message: string;
+                };
+            };
+            userAttributes: {
+                getAttributes: {
                     description: string;
                     message: string;
                 };